refactor(typeahead): factor out common test functions
